# src/main/resources/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "contact" : {
      "name" : "PowerCards Support",
      "url" : "https://github.com/victobonetti/powercards"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "PowerCards API",
    "version" : "1.0.0"
  },
  "paths" : {
...
    "/v1/decks" : {
      "get" : {
        "summary" : "List all decks",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "perPage",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "search",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginatedResponseDeckResponse"
                }
              }
            }
          }
        },
        "tags" : [ "Deck Resource" ]
      },
      "post" : {
        "summary" : "Create a new deck",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeckRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Deck created"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Deck Resource" ]
      }
    },
    "/v1/decks/{id}" : {
      "put" : {
        "summary" : "Update an existing deck",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeckRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Deck updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeckResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Deck not found"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "tags" : [ "Deck Resource" ]
      },
      "get" : {
        "summary" : "Get a deck by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Deck found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeckResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Deck not found"
          }
        },
        "tags" : [ "Deck Resource" ]
      },
      "delete" : {
        "summary" : "Delete a deck",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deck deleted"
          },
          "404" : {
            "description" : "Deck not found"
          }
        },
        "tags" : [ "Deck Resource" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "DeckRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        }
      },
      "DeckResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "cardCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "PaginatedResponseDeckResponse" : {
        "type" : "object",
        "properties" : {
          "pagination" : {
            "$ref" : "#/components/schemas/PaginationMeta"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DeckResponse"
            }
          }
        }
      },
...
      "PaginationMeta" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nextPageUri" : {
            "type" : "string"
          },
          "lastPageUri" : {
            "type" : "string"
          }
        }
      },
...
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8088",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8088",
    "description" : "Auto generated value"
  } ]
}